requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-countries==7.3.2
django-phonenumber-field==6.3.0
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.51
platformdirs==2.5.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
